Question Class, Basic CLI
I have, I think, finished the Question class. It can support questions with a correct answer, and any number of incorrect answers. You can get a random set of the answer choices, and check a string to see if it is equal to the correct answer.
I made a basic CLI to test this, where it asks a question, and you enter a number of an answer choice.
Next up is the Quiz object.