Added Quiz Class, more CLI Functionality
I added the Quiz class with functionality for constructing with a category and a title and an empty question list, and the ability to add questions. The List can be gotten and used, and I made a neat little property for just getting the number of questions. I'm not sure exactly how useful it is since you can also just get the list and check its size, but I wanted to try out the shorthand get notation.
The setters for the category and the title are set to internal, since I want them to only be changed by the QuizManager but not by anything using the API. I know that the CLI can currently just change it if it wanted to since its in the same namespace and assembly, but I think if I have the time to make a frontend, it won't be able to, so that's kind of what I want.