RandomQuiz w/ API Usage and Better CLI
I got the Open Trivia DB Usage down, so now I can generate random quizzes on any of the Categories! It's very cool, and I had fun trying out the quizzes. I also made the CLI for taking the quizzes look better, and moved it to its own method. Next us is to try and get a QuizManager to hold all of the quizzes in their categories. Then, get them all in their own JSON files.