QuizManager, Much Better CLI
Made the QuizManager! I have a much better idea now of how something like a server or an API works, I figured it out, I was thinking about it kind of wrong. I really needed to fully separate the client and the server and say if I was a client, what information would I need to be able to deliver this quiz. Plus, working off of the structure of what the OpenTriviaDB API was already doing helped out.
QuizManager is the touching point for the interfaces, so the CLI just makes one and uses it to manage (literally) all of the quizzes.
I also changed the CLI around a lot to be able to use the QuizManager, and to be able to take multiple quizzes. One thing I still need to do is see if I might be able to enforce time limits. The fields are there, but as of right now they don't get used by the CLI.
The QuizManager was tricky. I was thinking a lot about how I might want to use a web API with this, so I tried to make as many things strings as I could. I didn't want to commit to having everything be a string and jsonify all the lists and strings already, that could be changed pretty easily in the future, it's just another step at the start and end of those functions. I also messed around a lot with making the code cleaner, if I wanted to throw exceptions, if I wanted to get rid of empty CategoryLists, how to deal with that when editing (deleting then adding back), etc. Lots of little things that took up time. It's a very different workflow from the project I did last summer in C, but I still find it very fun seeing it all come together. Also, I'm liking C#, it's fun. I'm learning a lot!
Also, I want to be able to save custom quizzes that get made to the file system so that they can be reloaded in by the manager as it gets created, and when any quiz is added or removed or edited, the files are changed with it. I just thought about it, I might have to switch some things around in the manager code to be able to tell when to delete a folder when it is out of custom quizzes, even if there is a random quiz in there. I will have to do some thinking on that, that is tricky.
I don't think I will be able to do a full React interface tomorrow (or today, I guess), so I will probably maybe have to stick with just the command line interface I have. It's not ideal, I was hoping to give react and making a web API a try, but it's okay. I'm happy with how this is turning out, it works well and is scalable, even if it's not finished and it, overall, isn't actually a lot of code. The structure and design takes a lot of work to figure out, even if the code isn't all that crazy, I think.