Web API, Frontend Start
Moved all the code to a project with a react frontend and a ASP.NET core backend with an API to interface with the React Frontend. I got the Web API totally up and running, fully functional! However, I don't have time to make the Frontend do anything but display a random quiz and show off some of the features in the code.
There is still the CLI version, but that has some older code, and I changed some of the Quiz and Question code from the CLI to the Web API version. It should still work, though.